converted the bill variable from int to float

+++ main.py
@@ -1,6 +1,6 @@
 print("Welcome to the tip calculator!")
 
-bill = int(input("What was the total bill? $"))
+bill = float(input("What was the total bill? $"))
 tip = int(input("How much tip would you like to give? 10, 12, or 15? ")) + 100
 num_of_people = int(input("How many people to split the bill? "))
